Unified TestInstance and TestEntry into TestElement, removed JSONParser class to avoid multiple redundant reading/writing
